requirements.txt
# Derivtex - Deriv Volatility 30 Trading Bot
# Core dependencies
websockets>=12.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Technical indicators
ta>=0.10.0  # or use custom implementation

# Monitoring & Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Notifications
python-telegram-bot>=20.0

# Utilities
click>=8.1.0
rich>=13.0.0
python-json-logger>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0